requirements/dev.txt
-r base.txt

black==21.11b0
isort==5.10.1
flake8==4.0.1
django-debug-toolbar==3.2.2
ipdb==0.13.9
django-extensions==3.1.3
